requirements.txt
discord.py==1.7.3
jthon==1.0.2
aiosqlite==0.17.0
